refactor dahsboard into own view. rm maps for now
