+++ project/npda/views/__init__.py
@@ -4,3 +4,4 @@ from .npda_users import *
 from .patient import *
 from .submissions import *
 from .visit import *
+from .dashboard import dashboard
